CodeCov and moved EditAction to edit package
