+++ src/main/java/org/jabref/gui/maintable/RightClickMenu.java
@@ -7,7 +7,7 @@ import javafx.scene.control.SeparatorMenuItem;
 import org.jabref.Globals;
 import org.jabref.gui.BasePanel;
 import org.jabref.gui.DialogService;
-import org.jabref.gui.EditAction;
+import org.jabref.gui.edit.EditAction;
 import org.jabref.gui.SendAsEMailAction;
 import org.jabref.gui.StateManager;
 import org.jabref.gui.actions.ActionFactory;
